refactor: refactor the frontend code into a modular design instead of monolithic files we had (`index.html`, `app.css` and `app.js`)
