+++ internal/web/static/css/components/header.css
@@ -1,0 +1,144 @@
+/* Header - Top navigation bar styles */
+
+.main-header {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  padding: 0.5rem 1rem;
+  background: var(--pico-background-color);
+  border-bottom: 1px solid var(--pico-muted-border-color);
+  position: sticky;
+  top: 0;
+  z-index: 50;
+}
+
+.search-container {
+  flex: 1;
+  max-width: 500px;
+}
+
+.search-input {
+  margin: 0;
+  padding: 0.5rem 1rem 0.5rem 2.5rem;
+  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='18' height='18' viewBox='0 0 24 24' fill='none' stroke='%2394a3b8' stroke-width='2' stroke-linecap='round' stroke-linejoin='round'%3E%3Ccircle cx='11' cy='11' r='8'%3E%3C/circle%3E%3Cline x1='21' y1='21' x2='16.65' y2='16.65'%3E%3C/line%3E%3C/svg%3E");
+  background-repeat: no-repeat;
+  background-position: 0.75rem center;
+}
+
+.header-actions {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.header-actions-right {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  margin-left: auto;
+}
+
+.view-toggle {
+  display: flex;
+  gap: 0.125rem;
+  background: var(--pico-muted-border-color);
+  border-radius: 0.25rem;
+  padding: 0.125rem;
+}
+
+.view-toggle .btn-icon {
+  padding: 0.25rem;
+  border-radius: 0.2rem;
+}
+
+.view-toggle .btn-icon.active {
+  background: var(--snipo-primary);
+  color: white;
+}
+
+.view-toggle .btn-icon svg {
+  width: 14px;
+  height: 14px;
+}
+
+/* Draft recovery banner */
+.draft-banner {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0.75rem 1.5rem;
+  background: var(--snipo-warning-bg, #fef3c7);
+  border-bottom: 1px solid var(--snipo-warning-border, #f59e0b);
+  gap: 1rem;
+  flex-wrap: wrap;
+}
+
+[data-theme="dark"] .draft-banner {
+  background: rgba(245, 158, 11, 0.15);
+  border-color: rgba(245, 158, 11, 0.3);
+}
+
+.draft-banner-content {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+  color: #92400e;
+  font-size: 0.875rem;
+}
+
+[data-theme="dark"] .draft-banner-content {
+  color: #fbbf24;
+}
+
+.draft-banner-content svg {
+  flex-shrink: 0;
+}
+
+.draft-banner-actions {
+  display: flex;
+  gap: 0.5rem;
+  justify-content: flex-end;
+}
+
+.draft-banner-actions button {
+  padding: 0.4rem 0.75rem;
+  font-size: 0.85rem;
+}
+
+.draft-item-title {
+  font-weight: 500;
+  font-size: 0.85rem;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.draft-item-time {
+  font-size: 0.75rem;
+  color: var(--pico-muted-color);
+}
+
+.draft-item-actions {
+  display: flex;
+  gap: 0.5rem;
+  flex-shrink: 0;
+}
+
+.draft-item-title {
+  font-weight: 500;
+  font-size: 0.85rem;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.draft-item-time {
+  font-size: 0.75rem;
+  color: var(--pico-muted-color);
+}
+
+.draft-item-actions {
+  display: flex;
+  gap: 0.5rem;
+  flex-shrink: 0;
+}
