+++ internal/web/static/css/components/snippets.css
@@ -1,0 +1,148 @@
+/* Snippets - Snippet list and card styles */
+
+.snippet-list {
+  padding: 1.5rem;
+}
+
+.snippet-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
+  gap: 0.75rem;
+}
+
+/* List view */
+.snippet-list-view {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.snippet-list-view .snippet-card {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  padding: 0.5rem 0.75rem;
+}
+
+.snippet-list-view .snippet-card-header {
+  flex: 1;
+  min-width: 0;
+  margin-bottom: 0;
+}
+
+.snippet-list-view .snippet-card-description,
+.snippet-list-view .snippet-card-preview {
+  display: none;
+}
+
+.snippet-list-view .snippet-card-footer {
+  flex-shrink: 0;
+}
+
+.snippet-list-view .snippet-card-tags {
+  display: none;
+}
+
+.snippet-card {
+  background: var(--pico-card-background-color);
+  border: 1px solid var(--pico-muted-border-color);
+  border-radius: 0.375rem;
+  padding: 0.75rem;
+  cursor: pointer;
+  transition: border-color 0.15s, box-shadow 0.15s;
+}
+
+.snippet-card:hover {
+  border-color: var(--snipo-primary);
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+.snippet-card-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 0.375rem;
+  gap: 0.5rem;
+}
+
+.snippet-card-title {
+  font-weight: 500;
+  font-size: 0.7rem;
+  margin: 0;
+  color: var(--pico-color);
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  flex: 1;
+  min-width: 0;
+}
+
+.snippet-card-language {
+  font-size: 0.6rem;
+  text-transform: uppercase;
+  background: var(--snipo-primary);
+  color: white;
+  padding: 0.1rem 0.35rem;
+  border-radius: 0.2rem;
+  flex-shrink: 0;
+}
+
+.snippet-card-description {
+  font-size: 0.7rem;
+  line-height: 1.4;
+  color: var(--pico-muted-color);
+  margin: 0 0 0.375rem 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  line-clamp: 2;
+  -webkit-box-orient: vertical;
+}
+
+.snippet-card-preview {
+  font-family: 'Fira Code', 'Monaco', 'Consolas', monospace;
+  font-size: 0.55rem;
+  line-height: 1.3;
+  background: var(--card-preview-bg, var(--editor-bg));
+  color: var(--editor-text, #e2e8f0);
+  padding: 0.4rem;
+  border-radius: 0.25rem;
+  overflow: hidden;
+  max-height: 40px;
+  margin-bottom: 0.375rem;
+  margin-top: 0;
+  white-space: pre;
+}
+
+.snippet-card-preview code {
+  background: transparent !important;
+  padding: 0 !important;
+  font-size: inherit;
+}
+
+.snippet-card-footer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  font-size: 0.65rem;
+  color: var(--pico-muted-color);
+}
+
+.snippet-card-tags {
+  display: flex;
+  gap: 0.25rem;
+  flex-wrap: wrap;
+}
+
+.snippet-card-meta {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+}
+
+@media (max-width: 768px) {
+  .snippet-grid {
+    grid-template-columns: 1fr;
+  }
+}
