+++ internal/web/static/css/main.css
@@ -1,0 +1,14 @@
+/* Main CSS - Imports all component stylesheets */
+
+/* Core styles */
+@import './components/variables.css';
+@import './components/layout.css';
+
+/* Component styles */
+@import './components/sidebar.css';
+@import './components/header.css';
+@import './components/snippets.css';
+@import './components/editor.css';
+@import './components/modals.css';
+@import './components/buttons.css';
+@import './components/utilities.css';
