--- internal/web/static/js/app.js
@@ -1,2059 +1,0 @@
-// Snipo - Alpine.js Application
-
-// Theme management
-const theme = {
-  get() {
-    return localStorage.getItem('snipo-theme') || 'dark';
-  },
-  set(value) {
-    localStorage.setItem('snipo-theme', value);
-    document.documentElement.setAttribute('data-theme', value);
-    // Update Prism theme
-    this.updatePrismTheme(value);
-    // Update Ace editor theme
-    this.updateAceTheme(value);
-  },
-  toggle() {
-    const current = this.get();
-    this.set(current === 'dark' ? 'light' : 'dark');
-  },
-  init() {
-    const saved = this.get();
-    document.documentElement.setAttribute('data-theme', saved);
-    this.updatePrismTheme(saved);
-  },
-  updatePrismTheme(themeName) {
-    const prismLink = document.getElementById('prism-theme');
-    if (prismLink) {
-      if (themeName === 'dark') {
-        prismLink.href = '/static/vendor/css/prism-tomorrow.min.css';
-      } else {
-        prismLink.href = '/static/vendor/css/prism.min.css';
-      }
-    }
-  },
-  updateAceTheme(themeName) {
-    // Find the Ace editor instance and update its theme
-    if (typeof ace !== 'undefined') {
-      const editors = document.querySelectorAll('.ace_editor');
-      editors.forEach(editorEl => {
-        const editor = ace.edit(editorEl);
-        if (editor) {
-          const aceTheme = themeName === 'dark' ? 'ace/theme/monokai' : 'ace/theme/chrome';
-          editor.setTheme(aceTheme);
-        }
-      });
-    }
-  }
-};
-
-// Initialize theme on load
-theme.init();
-
-// API helper
-const api = {
-  async request(method, url, data = null) {
-    const options = {
-      method,
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include'
-    };
-    if (data) options.body = JSON.stringify(data);
-
-    const response = await fetch(url, options);
-    if (response.status === 401) {
-      window.location.href = '/login';
-      return null;
-    }
-    if (response.status === 204) return null;
-    return response.json();
-  },
-
-  get: (url) => api.request('GET', url),
-  post: (url, data) => api.request('POST', url, data),
-  put: (url, data) => api.request('PUT', url, data),
-  delete: (url) => api.request('DELETE', url)
-};
-
-// Toast notifications
-function showToast(message, type = 'success') {
-  const container = document.getElementById('toast-container');
-  if (!container) return;
-
-  const toast = document.createElement('div');
-  toast.className = `toast ${type}`;
-  toast.innerHTML = `
-    <span>${message}</span>
-    <button onclick="this.parentElement.remove()" style="background:none;border:none;cursor:pointer;padding:0;margin-left:0.5rem;">
-      <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-        <line x1="18" y1="6" x2="6" y2="18"></line>
-        <line x1="6" y1="6" x2="18" y2="18"></line>
-      </svg>
-    </button>
-  `;
-  container.appendChild(toast);
-  setTimeout(() => toast.remove(), 5000);
-}
-
-// Main app store
-document.addEventListener('alpine:init', () => {
-  // Global app state
-  Alpine.store('app', {
-    sidebarOpen: window.innerWidth > 768,
-    currentView: 'snippets', // snippets, editor, settings
-    loading: false,
-    darkMode: theme.get() === 'dark',
-
-    toggleSidebar() {
-      this.sidebarOpen = !this.sidebarOpen;
-    },
-
-    toggleTheme() {
-      theme.toggle();
-      this.darkMode = theme.get() === 'dark';
-    }
-  });
-
-  // Snippets data
-  // Expose helpers globally to avoid Alpine scope issues
-  window.autoResizeInput = function (element) {
-    if (!element) return;
-    const val = element.value || element.placeholder || '';
-    const length = val.length;
-    element.style.width = Math.max(10, Math.ceil(length * 1.5)) + 'ch';
-  };
-
-  window.autoResizeSelect = function (element) {
-    if (!element) return;
-    const selectedOption = element.options[element.selectedIndex];
-    const text = selectedOption ? selectedOption.text : element.value;
-    const span = document.createElement('span');
-    span.style.font = window.getComputedStyle(element).font;
-    span.style.visibility = 'hidden';
-    span.style.position = 'absolute';
-    span.textContent = text;
-    document.body.appendChild(span);
-    const width = span.offsetWidth + 35;
-    document.body.removeChild(span);
-    element.style.width = width + 'px';
-  };
-
-  Alpine.data('snippetsApp', () => ({
-    snippets: [],
-    tags: [],
-    folders: [],
-    selectedSnippet: null,
-    // Initialize with empty object to prevent null access errors
-    editingSnippet: {
-      id: null,
-      title: '',
-      description: '',
-      content: '',
-      language: 'plaintext',
-      tags: [],
-      folder_id: null,
-      is_public: false,
-      is_favorite: false,
-      files: [{
-        id: 0,
-        filename: 'snippet.txt',
-        content: '',
-        language: 'plaintext'
-      }]
-    },
-    activeFileIndex: 0, // Currently active file tab
-    editorHeaderVisible: true, // Toggle for editor toolbar/header
-    
-    // File Manager State - Single source of truth for file operations
-    fileManagerState: {
-      operationInProgress: false,  // Lock flag to prevent concurrent operations
-      editorDirty: false,           // Track if editor has unsaved changes
-      lastSyncedContent: '',        // Last content synced from editor to file
-      pendingOperation: null        // Queue for operations during locks
-    },
-    
-    filter: {
-      query: '',
-      tagId: null,
-      folderId: null,
-      language: '',
-      isFavorite: null,
-      isArchived: null
-    },
-    pagination: { page: 1, limit: 20, total: 0, totalPages: 0 },
-    totalSnippets: 0, // Total count for "All Snippets" (unfiltered)
-    favoritesCount: 0, // Count of favorite snippets
-    loading: true,
-    viewMode: localStorage.getItem('snipo-view-mode') || 'grid', // 'grid' or 'list'
-    showEditor: false,
-    isEditing: false, // false = preview mode, true = edit mode
-    showDeleteModal: false,
-    deleteTarget: null,
-    showSettings: false,
-    settingsTab: 'password', // 'password' or 'apikeys'
-    apiTokens: [],
-    newToken: { name: '', permissions: 'read', expires_in_days: 30 },
-    createdToken: null, // Stores newly created token for display
-    passwordForm: { current: '', new: '', confirm: '' },
-    passwordError: '',
-    passwordSuccess: '',
-    // Sidebar collapse state
-    foldersCollapsed: false,
-    tagsCollapsed: false,
-    // Folder/Tag management
-    showFolderModal: false,
-    showTagModal: false,
-    editingFolder: { name: '', parent_id: '' },
-    editingTag: { id: null, name: '' },
-    // Draft auto-save (single draft only)
-    hasDraft: false,
-    draftSnippet: null,
-    draftSavedAt: null,
-    autoSaveTimeout: null,
-    // Backup state
-    backupOptions: { format: 'json', password: '' },
-    importOptions: { strategy: 'merge', password: '' },
-    backupFile: null,
-    backupLoading: false,
-    importResult: null,
-    s3Status: { enabled: false },
-    s3Backups: [],
-    settings: { archive_enabled: false, history_enabled: true }, // Application settings
-    // History state
-    showHistory: false,
-    history: [],
-    historyLoading: false,
-    expandedHistory: null,
-    showHistoryDetail: false,
-    viewingHistoryEntry: null,
-    showRestoreConfirm: false,
-    restoringHistoryEntry: null,
-    // Ace Editor instance
-    aceEditor: null,
-    aceIgnoreChange: false, // Flag to prevent infinite loops
-
-    async init() {
-      await Promise.all([
-        this.loadSnippets(),
-        this.loadTags(),
-        this.loadFolders(),
-        this.loadSnippets(),
-        this.loadTags(),
-        this.loadFolders(),
-        this.loadFavoritesCount(),
-        this.loadSettings()
-      ]);
-      // Store total snippets count on initial load
-      this.totalSnippets = this.pagination.total;
-      this.loading = false;
-      // Highlight code after initial load
-      this.$nextTick(() => this.highlightAll());
-
-      // Restore state from URL
-      await this.restoreFromUrl();
-
-      // Handle browser back/forward
-      window.addEventListener('popstate', () => this.restoreFromUrl());
-
-      // Load draft if exists
-      this.loadDraft();
-    },
-
-    // URL routing
-    updateUrl(params = {}) {
-      const url = new URL(window.location);
-
-      // Clear existing params
-      url.searchParams.delete('snippet');
-      url.searchParams.delete('edit');
-      url.searchParams.delete('folder');
-      url.searchParams.delete('tag');
-      url.searchParams.delete('favorites');
-
-      // Set new params
-      if (params.snippet) url.searchParams.set('snippet', params.snippet);
-      if (params.edit) url.searchParams.set('edit', 'true');
-      if (params.folder) url.searchParams.set('folder', params.folder);
-      if (params.tag) url.searchParams.set('tag', params.tag);
-      if (params.favorites) url.searchParams.set('favorites', 'true');
-
-      history.pushState({}, '', url);
-    },
-
-    async restoreFromUrl() {
-      const params = new URLSearchParams(window.location.search);
-
-      const snippetId = params.get('snippet');
-      const isEdit = params.get('edit') === 'true';
-      const folderId = params.get('folder');
-      const tagId = params.get('tag');
-      const favorites = params.get('favorites') === 'true';
-
-      // Restore filter state
-      if (folderId) {
-        this.filter.folderId = parseInt(folderId);
-        this.filter.tagId = null;
-        this.filter.isFavorite = null;
-        await this.loadSnippets();
-      } else if (tagId) {
-        this.filter.tagId = parseInt(tagId);
-        this.filter.folderId = null;
-        this.filter.isFavorite = null;
-        await this.loadSnippets();
-      } else if (favorites) {
-        this.filter.isFavorite = true;
-        this.filter.tagId = null;
-        this.filter.folderId = null;
-        await this.loadSnippets();
-      }
-
-      // Restore snippet view/edit state
-      if (snippetId) {
-        const result = await api.get(`/api/v1/snippets/${snippetId}`);
-        if (result && !result.error) {
-          this.editingSnippet = {
-            ...result,
-            tags: (result.tags || []).map(t => t.name),
-            folder_id: result.folders?.[0]?.id || null,
-            files: result.files || []
-          };
-          this.activeFileIndex = 0;
-          this.showEditor = true;
-          this.isEditing = isEdit;
-          this.$nextTick(() => {
-            if (isEdit) {
-              this.updateAceEditor();
-            }
-            this.highlightAll();
-          });
-        }
-      }
-    },
-
-    async loadFavoritesCount() {
-      const result = await api.get('/api/v1/snippets?favorite=true&limit=1');
-      if (result && result.pagination) {
-        this.favoritesCount = result.pagination.total;
-      }
-    },
-
-    highlightAll() {
-      // Re-run Prism highlighting on all code blocks
-      if (typeof Prism !== 'undefined') {
-        Prism.highlightAll();
-      }
-    },
-
-    renderMarkdown(content) {
-      // Render markdown content as HTML using marked.js
-      if (!content) return '';
-      if (typeof marked !== 'undefined') {
-        // Configure marked for safe rendering
-        marked.setOptions({
-          breaks: true,
-          gfm: true
-        });
-        return marked.parse(content);
-      }
-      // Fallback: return content as-is if marked is not available
-      return content;
-    },
-
-    // Ace Editor language mode mapping
-    getAceMode(language) {
-      const modeMap = {
-        'javascript': 'ace/mode/javascript',
-        'typescript': 'ace/mode/typescript',
-        'python': 'ace/mode/python',
-        'go': 'ace/mode/golang',
-        'rust': 'ace/mode/rust',
-        'java': 'ace/mode/java',
-        'csharp': 'ace/mode/csharp',
-        'cpp': 'ace/mode/c_cpp',
-        'cuda': 'ace/mode/cuda',
-        'ruby': 'ace/mode/ruby',
-        'php': 'ace/mode/php',
-        'swift': 'ace/mode/swift',
-        'kotlin': 'ace/mode/kotlin',
-        'html': 'ace/mode/html',
-        'css': 'ace/mode/css',
-        'sql': 'ace/mode/sql',
-        'bash': 'ace/mode/sh',
-        'json': 'ace/mode/json',
-        'yaml': 'ace/mode/yaml',
-        'markdown': 'ace/mode/markdown',
-        'plaintext': 'ace/mode/text'
-      };
-      return modeMap[language] || 'ace/mode/text';
-    },
-
-    // Initialize or update Ace Editor
-    updateAceEditor() {
-      const container = this.$refs.codeEditor;
-      if (!container) return;
-
-      // Get current content and language
-      const content = (this.editingSnippet.files && this.editingSnippet.files.length > 0)
-        ? (this.activeFile?.content || '')
-        : (this.editingSnippet.content || '');
-
-      const language = (this.editingSnippet.files && this.editingSnippet.files.length > 0)
-        ? (this.activeFile?.language || 'javascript')
-        : (this.editingSnippet.language || 'javascript');
-
-      const aceMode = this.getAceMode(language);
-      const isDark = theme.get() === 'dark';
-      const aceTheme = isDark ? 'ace/theme/monokai' : 'ace/theme/chrome';
-
-      // If no editor exists, create one
-      if (!this.aceEditor) {
-        if (typeof ace === 'undefined') {
-          console.error('Ace Editor not loaded');
-          return;
-        }
-
-        try {
-          // Set base path for Ace to find modes and themes
-          ace.config.set('basePath', '/static/vendor/js/ace');
-
-          // Ensure container has an ID for Ace
-          if (!container.id) {
-            container.id = 'ace-editor-' + Date.now();
-          }
-
-          this.aceEditor = ace.edit(container.id);
-          this.aceEditor.setTheme(aceTheme);
-          this.aceEditor.session.setMode(aceMode);
-          this.aceEditor.setValue(content, -1); // -1 moves cursor to start
-
-          // Configure editor
-          this.aceEditor.setOptions({
-            fontSize: '14px',
-            showPrintMargin: false,
-            tabSize: 2,
-            useSoftTabs: true,
-            wrap: true
-          });
-
-          // Handle changes
-          const self = this;
-          this.aceEditor.session.on('change', () => {
-            if (self.aceIgnoreChange) return;
-
-            const value = self.aceEditor.getValue();
-            if (self.editingSnippet.files && self.editingSnippet.files.length > 0) {
-              self.updateActiveFileContent(value);
-            } else {
-              self.editingSnippet.content = value;
-              self.scheduleAutoSave();
-            }
-          });
-        } catch (e) {
-          console.error('Ace Editor initialization error:', e);
-          this.aceEditor = null;
-          return;
-        }
-      } else {
-        // Editor exists - update content and mode
-        this.aceIgnoreChange = true;
-        try {
-          // Always update the content to ensure it's in sync
-          // Don't rely on the equality check as it can miss reactive updates
-          this.aceEditor.setValue(content, -1);
-          this.aceEditor.session.setMode(aceMode);
-          this.aceEditor.setTheme(aceTheme);
-        } catch (e) {
-          console.warn('Ace Editor update error:', e);
-        }
-        this.aceIgnoreChange = false;
-      }
-
-      // Resize after DOM update
-      this.$nextTick(() => {
-        if (this.aceEditor) {
-          this.aceEditor.resize();
-        }
-      });
-    },
-
-    // Destroy Ace Editor instance
-    destroyAceEditor() {
-      if (this.aceEditor) {
-        try {
-          this.aceEditor.destroy();
-          // Clear the container
-          const container = this.$refs.codeEditor;
-          if (container) {
-            container.innerHTML = '';
-            container.className = 'ace-editor-container';
-          }
-        } catch (e) {
-          // Ignore errors during cleanup
-        }
-        this.aceEditor = null;
-      }
-    },
-
-    setViewMode(mode) {
-      this.viewMode = mode;
-      localStorage.setItem('snipo-view-mode', mode);
-    },
-
-    async loadSnippets() {
-      const params = new URLSearchParams();
-      params.set('page', this.pagination.page);
-      params.set('limit', this.pagination.limit);
-      if (this.filter.query) params.set('q', this.filter.query);
-      if (this.filter.tagId) params.set('tag_id', this.filter.tagId);
-      if (this.filter.folderId) params.set('folder_id', this.filter.folderId);
-      if (this.filter.language) params.set('language', this.filter.language);
-      if (this.filter.isFavorite !== null) params.set('favorite', this.filter.isFavorite);
-      if (this.filter.isArchived !== null) params.set('is_archived', this.filter.isArchived);
-
-      const result = await api.get(`/api/v1/snippets?${params}`);
-      if (result) {
-        this.snippets = result.data || [];
-        this.pagination = result.pagination || this.pagination;
-        // Highlight code after snippets load
-        this.$nextTick(() => this.highlightAll());
-      }
-    },
-
-    async loadTags() {
-      const result = await api.get('/api/v1/tags');
-      if (result) this.tags = result.data || [];
-    },
-
-    async loadFolders() {
-      const result = await api.get('/api/v1/folders?tree=true');
-      if (result) this.folders = result.data || [];
-    },
-
-    async loadSettings() {
-      const result = await api.get('/api/v1/settings');
-      if (result) {
-        this.settings = result;
-      }
-    },
-
-    async updateSettings() {
-      const result = await api.put('/api/v1/settings', this.settings);
-      if (result) {
-        this.settings = result;
-        showToast('Settings updated');
-      }
-    },
-
-    async search() {
-      this.pagination.page = 1;
-      await this.loadSnippets();
-    },
-
-    async filterByTag(tagId) {
-      this.showEditor = false;
-      this.filter.tagId = tagId;
-      this.filter.folderId = null;
-      this.filter.isFavorite = null;
-      this.pagination.page = 1;
-      await this.loadSnippets();
-      this.updateUrl({ tag: tagId });
-    },
-
-    async filterByFolder(folderId) {
-      this.showEditor = false;
-      this.filter.folderId = folderId;
-      this.filter.tagId = null;
-      this.filter.isFavorite = null;
-      this.pagination.page = 1;
-      await this.loadSnippets();
-      this.updateUrl({ folder: folderId });
-    },
-
-    async clearFilters() {
-      this.showEditor = false;
-      this.filter = { query: '', tagId: null, folderId: null, language: '', isFavorite: null, isArchived: null };
-      this.pagination.page = 1;
-      await this.loadSnippets();
-      this.totalSnippets = this.pagination.total;
-      this.updateUrl({});
-    },
-
-    async showFavorites() {
-      this.showEditor = false;
-      this.filter.isFavorite = true;
-      this.filter.tagId = null;
-      this.filter.folderId = null;
-      this.pagination.page = 1;
-      await this.loadSnippets();
-      this.updateUrl({ favorites: true });
-    },
-
-    async showArchive() {
-      this.showEditor = false;
-      this.filter.isArchived = true;
-      this.filter.tagId = null;
-      this.filter.folderId = null;
-      this.filter.isFavorite = null;
-      this.pagination.page = 1;
-      await this.loadSnippets();
-      // URL update for archive? maybe ?archive=true
-    },
-
-    async toggleArchive(snippet) {
-      const result = await api.post(`/api/v1/snippets/${snippet.id}/archive`);
-      if (result) {
-        // If we are in detail view, update the snippet
-        if (this.editingSnippet.id === snippet.id) {
-          this.editingSnippet.is_archived = result.is_archived;
-        }
-
-        // If we filter by archive state, remove it from list
-        if (this.filter.isArchived !== null) {
-          this.snippets = this.snippets.filter(s => s.id !== snippet.id);
-        } else {
-          // Update in list
-          const idx = this.snippets.findIndex(s => s.id === snippet.id);
-          if (idx !== -1) {
-            this.snippets[idx].is_archived = result.is_archived;
-            // If we are showing "All" (default), archived snippets should be hidden?
-            // Backend default is hidden. If filter.isArchived is null (default),
-            // then getting the list again would hide it.
-            // But we want to avoid reload. 
-            // If is_archived became true, and filter is null (default = active only), remove it.
-            if (result.is_archived && this.filter.isArchived === null) {
-              this.snippets.splice(idx, 1);
-            }
-          }
-        }
-
-        showToast(result.is_archived ? 'Snippet archived' : 'Snippet unarchived');
-      }
-    },
-
-    newSnippet() {
-      this.editingSnippet = {
-        id: null,
-        title: '',
-        description: '',
-        content: '',
-        language: 'plaintext',
-        tags: [],
-        folder_id: null,
-        is_public: false,
-        is_favorite: false,
-        files: [{
-          id: 0,
-          filename: 'snippet.txt',
-          content: '',
-          language: 'plaintext'
-        }]
-      };
-      this.activeFileIndex = 0;
-      this.showEditor = true;
-      this.isEditing = true;
-      this.updateUrl({ edit: true });
-
-      // Update Ace Editor and focus filename input after render
-      this.$nextTick(() => {
-        this.updateAceEditor();
-        const input = document.querySelector('.filename-input');
-        if (input) {
-          input.focus();
-          input.select();
-        }
-      });
-    },
-
-    async viewSnippet(snippet) {
-      const result = await api.get(`/api/v1/snippets/${snippet.id}`);
-      if (result) {
-        this.editingSnippet = {
-          ...result,
-          tags: (result.tags || []).map(t => t.name),
-          folder_id: result.folders?.[0]?.id || null,
-          files: result.files || []
-        };
-        this.activeFileIndex = 0;
-        this.showEditor = true;
-        this.isEditing = false;
-        this.updateUrl({ snippet: snippet.id });
-        this.$nextTick(() => this.highlightAll());
-      }
-    },
-
-    startEditing() {
-      this.isEditing = true;
-      if (this.editingSnippet?.id) {
-        this.updateUrl({ snippet: this.editingSnippet.id, edit: true });
-      }
-      this.$nextTick(() => {
-        this.updateAceEditor();
-        this.highlightAll();
-      });
-    },
-
-    cancelEditing() {
-      // Switch back to preview mode without saving
-      this.isEditing = false;
-      if (this.editingSnippet?.id) {
-        // Reload the snippet to discard changes
-        this.viewSnippet(this.editingSnippet);
-      } else {
-        // If it's a new snippet, close the editor
-        this.showEditor = false;
-        this.editingSnippet = null;
-        this.updateUrl({ snippet: null, edit: null });
-      }
-    },
-
-    async editSnippet(snippet) {
-      const result = await api.get(`/api/v1/snippets/${snippet.id}`);
-      if (result) {
-        this.editingSnippet = {
-          ...result,
-          tags: (result.tags || []).map(t => t.name),
-          folder_id: result.folders?.[0]?.id || null,
-          files: result.files || []
-        };
-        this.activeFileIndex = 0;
-        this.showEditor = true;
-        this.isEditing = true;
-        this.updateUrl({ snippet: snippet.id, edit: true });
-        this.$nextTick(() => {
-          this.updateAceEditor();
-          this.highlightAll();
-        });
-      }
-    },
-
-    async saveSnippet() {
-      // Convert empty string folder_id to null, and string numbers to integers
-      let folderId = this.editingSnippet.folder_id;
-      if (folderId === '' || folderId === null || folderId === undefined) {
-        folderId = null;
-      } else {
-        folderId = parseInt(folderId, 10) || null;
-      }
-
-      // Prepare files array if multi-file
-      let files = null;
-      if (this.editingSnippet.files && this.editingSnippet.files.length > 0) {
-        files = this.editingSnippet.files.map(f => ({
-          id: f.id || 0,
-          filename: f.filename,
-          content: f.content,
-          language: f.language
-        }));
-      }
-
-      // For multi-file snippets, use first file's content/language as primary
-      const primaryContent = files && files.length > 0 ? files[0].content : this.editingSnippet.content;
-      const primaryLanguage = files && files.length > 0 ? files[0].language : this.editingSnippet.language;
-
-      const data = {
-        title: this.editingSnippet.title,
-        description: this.editingSnippet.description || '',
-        content: primaryContent,
-        language: primaryLanguage,
-        tags: this.editingSnippet.tags || [],
-        folder_id: folderId,
-        is_public: this.editingSnippet.is_public || false,
-        files: files
-      };
-
-      let result;
-      if (this.editingSnippet.id) {
-        result = await api.put(`/api/v1/snippets/${this.editingSnippet.id}`, data);
-      } else {
-        result = await api.post('/api/v1/snippets', data);
-      }
-
-      if (result && !result.error) {
-        showToast(this.editingSnippet.id ? 'Snippet updated' : 'Snippet created');
-        this.showEditor = false;
-        this.isEditing = false;
-        // Destroy Ace Editor when leaving editor
-        this.destroyAceEditor();
-        this.resetEditingSnippet();
-        this.clearDraft(); // Clear draft on successful save
-        this.updateUrl({}); // Clear URL params
-        await this.loadSnippets();
-        await this.loadTags();
-        await this.loadFavoritesCount();
-      } else if (result?.error) {
-        showToast(result.error.message || 'Error saving snippet', 'error');
-      }
-    },
-
-    cancelEdit() {
-      this.showEditor = false;
-      this.isEditing = false;
-      // Destroy Ace Editor when leaving editor
-      this.destroyAceEditor();
-      this.resetEditingSnippet();
-      this.clearDraft();
-      // Restore URL to current filter state
-      if (this.filter.folderId) {
-        this.updateUrl({ folder: this.filter.folderId });
-      } else if (this.filter.tagId) {
-        this.updateUrl({ tag: this.filter.tagId });
-      } else if (this.filter.isFavorite) {
-        this.updateUrl({ favorites: true });
-      } else {
-        this.updateUrl({});
-      }
-    },
-
-    resetEditingSnippet() {
-      this.editingSnippet = {
-        id: null,
-        title: '',
-        description: '',
-        content: '',
-        language: 'plaintext',
-        tags: [],
-        folder_id: null,
-        is_public: false,
-        is_favorite: false,
-        files: [{
-          id: 0,
-          filename: 'snippet.txt',
-          content: '',
-          language: 'plaintext'
-        }]
-      };
-      this.activeFileIndex = 0;
-    },
-
-    // Draft auto-save functionality (single draft only - always latest)
-    saveDraft() {
-      if (!this.isEditing) return;
-
-      // Check if there's content worth saving
-      const hasContent = this.editingSnippet.title ||
-        this.editingSnippet.content ||
-        (this.editingSnippet.files && this.editingSnippet.files.some(f => f.content));
-      if (!hasContent) return;
-
-      const draft = {
-        snippet: { ...this.editingSnippet },
-        savedAt: new Date().toISOString()
-      };
-      localStorage.setItem('snipo-draft', JSON.stringify(draft));
-    },
-
-    loadDraft() {
-      // Only show if we're not already viewing a snippet from URL
-      if (this.showEditor) return;
-
-      try {
-        const draftJson = localStorage.getItem('snipo-draft');
-        if (!draftJson) return;
-
-        const draft = JSON.parse(draftJson);
-        if (!draft.snippet) return;
-
-        // Check if draft is less than 24 hours old
-        const savedAt = new Date(draft.savedAt);
-        const now = new Date();
-        const hoursDiff = (now - savedAt) / (1000 * 60 * 60);
-
-        if (hoursDiff > 24) {
-          this.clearDraft();
-          return;
-        }
-
-        // Check if draft has content
-        const hasContent = draft.snippet.title ||
-          draft.snippet.content ||
-          (draft.snippet.files && draft.snippet.files.some(f => f.content));
-        if (hasContent) {
-          this.hasDraft = true;
-          this.draftSnippet = draft.snippet;
-          this.draftSavedAt = savedAt;
-        }
-      } catch (e) {
-        this.clearDraft();
-      }
-    },
-
-    restoreDraft() {
-      if (this.draftSnippet) {
-        this.editingSnippet = { ...this.draftSnippet };
-        this.activeFileIndex = 0;
-        this.showEditor = true;
-        this.isEditing = true;
-        this.hasDraft = false;
-        this.clearDraft();
-        showToast('Draft restored');
-        this.$nextTick(() => {
-          this.updateAceEditor();
-          this.highlightAll();
-        });
-      }
-    },
-
-    discardDraft() {
-      this.clearDraft();
-      this.hasDraft = false;
-      this.draftSnippet = null;
-      showToast('Draft discarded');
-    },
-
-    clearDraft() {
-      localStorage.removeItem('snipo-draft');
-    },
-
-    // Auto-save on content change (debounced)
-    scheduleAutoSave() {
-      if (this.autoSaveTimeout) {
-        clearTimeout(this.autoSaveTimeout);
-      }
-      this.autoSaveTimeout = setTimeout(() => {
-        this.saveDraft();
-      }, 2000); // Save after 2 seconds of inactivity
-    },
-
-    confirmDelete(snippet) {
-      this.deleteTarget = snippet;
-      this.showDeleteModal = true;
-    },
-
-    async deleteSnippet() {
-      if (!this.deleteTarget) return;
-
-      await api.delete(`/api/v1/snippets/${this.deleteTarget.id}`);
-      showToast('Snippet deleted');
-      this.showDeleteModal = false;
-      this.showEditor = false;
-      this.deleteTarget = null;
-
-      // Reload all data to update counts
-      await Promise.all([
-        this.loadSnippets(),
-        this.loadTags(),
-        this.loadFolders(),
-        this.loadFavoritesCount()
-      ]);
-
-      // Update total count
-      this.totalSnippets = this.snippets.length;
-    },
-
-    async toggleFavorite(snippet) {
-      const result = await api.post(`/api/v1/snippets/${snippet.id}/favorite`);
-      if (result) {
-        snippet.is_favorite = result.is_favorite;
-        showToast(result.is_favorite ? 'Added to favorites' : 'Removed from favorites');
-      }
-    },
-
-    async duplicateSnippet(snippet) {
-      const result = await api.post(`/api/v1/snippets/${snippet.id}/duplicate`);
-      if (result && !result.error) {
-        showToast('Snippet duplicated');
-        await this.loadSnippets();
-      }
-    },
-
-    async copyToClipboard(snippet) {
-      try {
-        await navigator.clipboard.writeText(snippet.content);
-        showToast('Copied to clipboard');
-      } catch (err) {
-        showToast('Failed to copy', 'error');
-      }
-    },
-
-    async copyShareLink(snippet) {
-      if (!snippet?.id || !snippet?.is_public) {
-        showToast('Snippet must be public to share', 'warning');
-        return;
-      }
-      try {
-        const shareUrl = `${window.location.origin}/s/${snippet.id}`;
-        await navigator.clipboard.writeText(shareUrl);
-        showToast('Share link copied to clipboard');
-      } catch (err) {
-        showToast('Failed to copy link', 'error');
-      }
-    },
-
-    // ============================================================================
-    // FILE MANAGER - Core file operation system
-    // ============================================================================
-    
-    /**
-     * Safely sync editor content to the current file
-     * This is the ONLY way content should be saved from editor to file
-     */
-    _syncEditorToFile() {
-      if (!this.aceEditor || !this.editingSnippet.files || !this.editingSnippet.files[this.activeFileIndex]) {
-        return;
-      }
-      
-      const currentContent = this.aceEditor.getValue();
-      this.editingSnippet.files[this.activeFileIndex].content = currentContent;
-      this.fileManagerState.lastSyncedContent = currentContent;
-      this.fileManagerState.editorDirty = false;
-    },
-
-    /**
-     * Safely load file content into the editor
-     * This is the ONLY way content should be loaded from file to editor
-     */
-    _loadFileToEditor(fileIndex) {
-      if (!this.aceEditor || !this.editingSnippet.files || !this.editingSnippet.files[fileIndex]) {
-        return;
-      }
-
-      const file = this.editingSnippet.files[fileIndex];
-      const content = file.content || '';
-      
-      // Disable change handler during load
-      this.aceIgnoreChange = true;
-      try {
-        this.aceEditor.setValue(content, -1);
-        this.aceEditor.session.setMode(this.getAceMode(file.language));
-        this.fileManagerState.lastSyncedContent = content;
-        this.fileManagerState.editorDirty = false;
-      } finally {
-        this.aceIgnoreChange = false;
-      }
-    },
-
-    /**
-     * Begin a file operation transaction
-     * Locks the file manager and syncs current state
-     */
-    _beginFileOperation() {
-      if (this.fileManagerState.operationInProgress) {
-        console.warn('File operation already in progress');
-        return false;
-      }
-      
-      // Lock the system
-      this.fileManagerState.operationInProgress = true;
-      
-      // Sync current editor content to file before any operation
-      this._syncEditorToFile();
-      
-      return true;
-    },
-
-    /**
-     * Complete a file operation transaction
-     * Unlocks the file manager and loads new state
-     */
-    _endFileOperation(newFileIndex) {
-      if (!this.fileManagerState.operationInProgress) {
-        return;
-      }
-
-      // Update active index
-      this.activeFileIndex = newFileIndex;
-      
-      // Load new file into editor
-      this.$nextTick(() => {
-        this._loadFileToEditor(newFileIndex);
-        
-        // Unlock the system
-        this.fileManagerState.operationInProgress = false;
-      });
-    },
-
-    // ============================================================================
-    // Multi-file support - Refactored to use File Manager
-    // ============================================================================
-    
-    get activeFile() {
-      const files = this.editingSnippet?.files || [];
-      if (files.length === 0) {
-        // Return legacy single-file as virtual file
-        return {
-          id: 0,
-          filename: 'main',
-          content: this.editingSnippet?.content || '',
-          language: this.editingSnippet?.language || 'javascript'
-        };
-      }
-      return files[this.activeFileIndex] || files[0];
-    },
-
-    get hasMultipleFiles() {
-      return (this.editingSnippet?.files || []).length > 1;
-    },
-
-    syncCurrentContent() {
-      // Legacy method - now delegates to file manager
-      this._syncEditorToFile();
-    },
-
-    addFile() {
-      // Begin transaction
-      if (!this._beginFileOperation()) {
-        return;
-      }
-
-      try {
-        if (!this.editingSnippet.files) {
-          // Convert legacy single-file to multi-file
-          // Content is already synced by _beginFileOperation
-          const ext = this.getFileExtension(this.editingSnippet.language);
-          this.editingSnippet.files = [{
-            id: 0,
-            filename: 'main.' + ext,
-            content: this.editingSnippet.content || '',
-            language: this.editingSnippet.language || 'javascript'
-          }];
-        }
-
-        // Add new file with placeholder name
-        const newFile = {
-          id: 0,
-          filename: 'newfile.txt',
-          content: '',
-          language: 'plaintext'
-        };
-        this.editingSnippet.files.push(newFile);
-        
-        const newIndex = this.editingSnippet.files.length - 1;
-        
-        // End transaction and switch to new file
-        this._endFileOperation(newIndex);
-
-        // Focus the filename input after render
-        setTimeout(() => {
-          const inputs = document.querySelectorAll('.filename-input');
-          if (inputs.length > 0) {
-            const lastInput = inputs[inputs.length - 1];
-            lastInput.focus();
-            lastInput.select();
-          }
-        }, 100);
-
-      } catch (error) {
-        console.error('Error adding file:', error);
-        this.fileManagerState.operationInProgress = false;
-      }
-
-      this.scheduleAutoSave();
-    },
-
-    // Auto-resize textarea for description
-    autoResizeTextarea(el) {
-      if (!el) return;
-      el.style.height = 'auto';
-      el.style.height = Math.min(el.scrollHeight, 80) + 'px';
-    },
-
-    validateFilename() {
-      if (this.editingSnippet.files && this.editingSnippet.files.length > 0) {
-        const file = this.editingSnippet.files[this.activeFileIndex];
-        if (!file.filename || !file.filename.trim()) {
-          file.filename = 'untitled.txt';
-          showToast('Filename cannot be empty', 'warning');
-        }
-      }
-    },
-
-    detectLanguageFromFilename(filename) {
-      const ext = filename.split('.').pop()?.toLowerCase();
-      const langMap = {
-        'js': 'javascript', 'ts': 'typescript', 'py': 'python', 'go': 'go',
-        'rs': 'rust', 'java': 'java', 'cs': 'csharp', 'cpp': 'cpp', 'c': 'cpp',
-        'cu': 'cuda', 'cuh': 'cuda',
-        'rb': 'ruby', 'php': 'php', 'swift': 'swift', 'kt': 'kotlin',
-        'html': 'html', 'css': 'css', 'sql': 'sql', 'sh': 'bash',
-        'json': 'json', 'yaml': 'yaml', 'yml': 'yaml', 'md': 'markdown',
-        'txt': 'plaintext'
-      };
-      return langMap[ext] || null;
-    },
-
-    removeFile(index) {
-      if (!this.editingSnippet.files || this.editingSnippet.files.length <= 1) {
-        showToast('Cannot remove the last file', 'warning');
-        return;
-      }
-
-      // Begin transaction
-      if (!this._beginFileOperation()) {
-        return;
-      }
-
-      try {
-        // Determine which file will become active after removal
-        let newActiveIndex;
-        if (index === this.activeFileIndex) {
-          // We're removing the active file, switch to the previous file or first if none before
-          newActiveIndex = Math.max(0, index - 1);
-        } else if (index < this.activeFileIndex) {
-          // If we're removing a file before the active one, shift index down
-          newActiveIndex = this.activeFileIndex - 1;
-        } else {
-          // Removing a file after the active one, index stays the same
-          newActiveIndex = this.activeFileIndex;
-        }
-
-        // Remove the file from array
-        this.editingSnippet.files.splice(index, 1);
-
-        // End transaction and switch to new active file
-        this._endFileOperation(newActiveIndex);
-
-      } catch (error) {
-        console.error('Error removing file:', error);
-        this.fileManagerState.operationInProgress = false;
-      }
-
-      this.scheduleAutoSave();
-    },
-
-    selectFile(index) {
-      // Begin transaction
-      if (!this._beginFileOperation()) {
-        return;
-      }
-
-      try {
-        // End transaction and switch to selected file
-        this._endFileOperation(index);
-        
-        // Highlight code after switch
-        this.$nextTick(() => {
-          this.highlightAll();
-        });
-      } catch (error) {
-        console.error('Error selecting file:', error);
-        this.fileManagerState.operationInProgress = false;
-      }
-    },
-
-    updateActiveFileContent(content) {
-      // This is called by the editor change handler
-      // Only update if not in the middle of an operation
-      if (this.fileManagerState.operationInProgress) {
-        return;
-      }
-
-      if (this.editingSnippet.files && this.editingSnippet.files.length > 0) {
-        this.editingSnippet.files[this.activeFileIndex].content = content;
-        this.fileManagerState.editorDirty = true;
-      } else {
-        this.editingSnippet.content = content;
-      }
-      this.scheduleAutoSave();
-    },
-
-    updateActiveFileLanguage(language) {
-      // Sync current content first
-      this._syncEditorToFile();
-      
-      // Get current content
-      const currentContent = this.aceEditor ? this.aceEditor.getValue() : '';
-
-      if (this.editingSnippet.files && this.editingSnippet.files.length > 0) {
-        this.editingSnippet.files[this.activeFileIndex].language = language;
-        // Also update content from editor
-        this.editingSnippet.files[this.activeFileIndex].content = currentContent;
-      } else {
-        this.editingSnippet.language = language;
-        this.editingSnippet.content = currentContent;
-      }
-
-      // Update Ace Editor mode
-      if (this.aceEditor) {
-        try {
-          this.aceEditor.session.setMode(this.getAceMode(language));
-        } catch (e) {
-          console.warn('Ace setMode error:', e);
-        }
-      }
-
-      this.$nextTick(() => this.highlightAll());
-      this.scheduleAutoSave();
-    },
-
-    updateActiveFilename(filename) {
-      if (this.editingSnippet.files && this.editingSnippet.files.length > 0) {
-        this.editingSnippet.files[this.activeFileIndex].filename = filename;
-        // Auto-detect language from extension
-        const detectedLang = this.detectLanguageFromFilename(filename);
-        if (detectedLang) {
-          const currentLang = this.editingSnippet.files[this.activeFileIndex].language;
-          if (detectedLang !== currentLang) {
-            this.editingSnippet.files[this.activeFileIndex].language = detectedLang;
-            // Update Ace Editor mode
-            if (this.aceEditor) {
-              try {
-                this.aceEditor.session.setMode(this.getAceMode(detectedLang));
-              } catch (e) {
-                console.warn('Ace setMode error:', e);
-              }
-            }
-          }
-        }
-      }
-      this.scheduleAutoSave();
-    },
-
-    getFileExtension(language) {
-      const extMap = {
-        'javascript': 'js', 'typescript': 'ts', 'python': 'py', 'go': 'go',
-        'rust': 'rs', 'java': 'java', 'csharp': 'cs', 'cpp': 'cpp', 'cuda': 'cu',
-        'ruby': 'rb', 'php': 'php', 'swift': 'swift', 'kotlin': 'kt',
-        'html': 'html', 'css': 'css', 'sql': 'sql', 'bash': 'sh',
-        'json': 'json', 'yaml': 'yaml', 'markdown': 'md', 'plaintext': 'txt'
-      };
-      return extMap[language] || 'txt';
-    },
-
-    formatDate(dateStr) {
-      const date = new Date(dateStr);
-      const now = new Date();
-      const diff = now - date;
-
-      if (diff < 60000) return 'Just now';
-      if (diff < 3600000) return `${Math.floor(diff / 60000)}m ago`;
-      if (diff < 86400000) return `${Math.floor(diff / 3600000)}h ago`;
-      if (diff < 604800000) return `${Math.floor(diff / 86400000)}d ago`;
-
-      return date.toLocaleDateString();
-    },
-
-    getLanguageColor(lang) {
-      const colors = {
-        javascript: '#f7df1e',
-        typescript: '#3178c6',
-        python: '#3776ab',
-        go: '#00add8',
-        rust: '#dea584',
-        java: '#b07219',
-        csharp: '#178600',
-        cpp: '#f34b7d',
-        cuda: '#76b900',
-        ruby: '#cc342d',
-        php: '#4f5d95',
-        swift: '#fa7343',
-        kotlin: '#a97bff',
-        html: '#e34c26',
-        css: '#563d7c',
-        sql: '#e38c00',
-        bash: '#89e051',
-        json: '#292929',
-        yaml: '#cb171e',
-        markdown: '#083fa1',
-        plaintext: '#6b7280'
-      };
-      return colors[lang] || '#6b7280';
-    },
-
-    addTag(tag) {
-      if (tag && !this.editingSnippet.tags.includes(tag)) {
-        this.editingSnippet.tags.push(tag);
-      }
-    },
-
-    removeTag(index) {
-      this.editingSnippet.tags.splice(index, 1);
-    },
-
-    async logout() {
-      await api.post('/api/v1/auth/logout');
-      window.location.href = '/login';
-    },
-
-    // Settings functions
-    async openSettings() {
-      this.showSettings = true;
-      this.settingsTab = 'password';
-      this.passwordForm = { current: '', new: '', confirm: '' };
-      this.passwordError = '';
-      this.passwordSuccess = '';
-      this.createdToken = null;
-      await this.loadApiTokens();
-    },
-
-    closeSettings() {
-      this.showSettings = false;
-      this.createdToken = null;
-    },
-
-    async loadApiTokens() {
-      const result = await api.get('/api/v1/tokens');
-      if (result && result.data) {
-        this.apiTokens = result.data;
-      }
-    },
-
-    async changePassword() {
-      this.passwordError = '';
-      this.passwordSuccess = '';
-
-      if (this.passwordForm.new !== this.passwordForm.confirm) {
-        this.passwordError = 'New passwords do not match';
-        return;
-      }
-
-      if (this.passwordForm.new.length < 6) {
-        this.passwordError = 'Password must be at least 6 characters';
-        return;
-      }
-
-      const result = await api.post('/api/v1/auth/change-password', {
-        current_password: this.passwordForm.current,
-        new_password: this.passwordForm.new
-      });
-
-      if (result && !result.error) {
-        this.passwordSuccess = 'Password changed successfully. Logging out...';
-        this.passwordForm = { current: '', new: '', confirm: '' };
-        // Logout after successful password change
-        setTimeout(async () => {
-          await this.logout();
-        }, 1500);
-      } else {
-        this.passwordError = result?.error?.message || 'Failed to change password';
-      }
-    },
-
-    async createApiToken() {
-      if (!this.newToken.name.trim()) {
-        showToast('Token name is required', 'error');
-        return;
-      }
-
-      const result = await api.post('/api/v1/tokens', {
-        name: this.newToken.name,
-        permissions: this.newToken.permissions,
-        expires_in_days: parseInt(this.newToken.expires_in_days) || null
-      });
-
-      if (result && !result.error) {
-        this.createdToken = result.token;
-        this.newToken = { name: '', permissions: 'read', expires_in_days: 30 };
-        await this.loadApiTokens();
-        showToast('API token created');
-      } else {
-        showToast(result?.error?.message || 'Failed to create token', 'error');
-      }
-    },
-
-    async deleteApiToken(tokenId) {
-      if (!confirm('Are you sure you want to delete this API token?')) return;
-
-      const result = await api.delete(`/api/v1/tokens/${tokenId}`);
-      if (!result || !result.error) {
-        await this.loadApiTokens();
-        showToast('API token deleted');
-      } else {
-        showToast('Failed to delete token', 'error');
-      }
-    },
-
-    copyToken() {
-      if (this.createdToken) {
-        navigator.clipboard.writeText(this.createdToken);
-        showToast('Token copied to clipboard');
-      }
-    },
-
-    // History functions
-    async openHistory(snippet) {
-      if (!snippet || !snippet.id) return;
-      
-      this.showHistory = true;
-      this.history = [];
-      this.historyLoading = true;
-      this.expandedHistory = null;
-
-      try {
-        const result = await api.get(`/api/v1/snippets/${snippet.id}/history?limit=50`);
-        if (result && result.data) {
-          this.history = result.data;
-        }
-      } catch (error) {
-        console.error('Failed to load history:', error);
-        showToast('Failed to load history', 'error');
-      } finally {
-        this.historyLoading = false;
-      }
-    },
-
-    closeHistory() {
-      this.showHistory = false;
-      this.history = [];
-      this.expandedHistory = null;
-    },
-
-    viewHistoryDetail(historyEntry) {
-      if (!historyEntry) return;
-      this.viewingHistoryEntry = { ...historyEntry, is_current: this.history.indexOf(historyEntry) === 0 };
-      this.showHistoryDetail = true;
-    },
-
-    closeHistoryDetail() {
-      this.showHistoryDetail = false;
-      this.viewingHistoryEntry = null;
-    },
-
-    copyHistoryContent(content) {
-      if (!content) return;
-      navigator.clipboard.writeText(content)
-        .then(() => showToast('Content copied to clipboard'))
-        .catch(() => showToast('Failed to copy content', 'error'));
-    },
-
-    confirmRestoreHistory(historyEntry) {
-      if (!historyEntry || !historyEntry.snippet_id || !historyEntry.id) {
-        showToast('Invalid history entry', 'error');
-        return;
-      }
-
-      // Check if trying to restore the current version
-      if (this.history.indexOf(historyEntry) === 0) {
-        showToast('This is already the current version', 'info');
-        return;
-      }
-
-      this.restoringHistoryEntry = historyEntry;
-      this.showRestoreConfirm = true;
-    },
-
-    async executeRestore() {
-      const historyEntry = this.restoringHistoryEntry;
-      if (!historyEntry) return;
-
-      // Validate snippet still exists
-      if (this.editingSnippet && this.editingSnippet.id !== historyEntry.snippet_id) {
-        showToast('Cannot restore: snippet mismatch', 'error');
-        this.showRestoreConfirm = false;
-        return;
-      }
-
-      this.showRestoreConfirm = false;
-
-      try {
-        const result = await api.post(
-          `/api/v1/snippets/${historyEntry.snippet_id}/history/${historyEntry.id}/restore`
-        );
-        
-        if (result) {
-          showToast('Snippet successfully restored from history');
-          
-          // Close all history modals
-          this.closeHistoryDetail();
-          this.closeHistory();
-          
-          // Reload the snippet to show restored version
-          if (this.editingSnippet && this.editingSnippet.id === historyEntry.snippet_id) {
-            await this.loadSnippet(historyEntry.snippet_id);
-          }
-          
-          // Reload snippet list
-          await this.loadSnippets();
-        } else {
-          throw new Error('No response from server');
-        }
-      } catch (error) {
-        console.error('Failed to restore history:', error);
-        
-        // Handle specific error cases
-        if (error.message && error.message.includes('404')) {
-          showToast('Snippet or history entry not found', 'error');
-        } else if (error.message && error.message.includes('403')) {
-          showToast('Permission denied', 'error');
-        } else if (error.message && error.message.includes('network')) {
-          showToast('Network error - please check your connection', 'error');
-        } else {
-          showToast('Failed to restore from history', 'error');
-        }
-      } finally {
-        this.restoringHistoryEntry = null;
-      }
-    },
-
-    toggleHistoryPreview(historyId) {
-      this.expandedHistory = this.expandedHistory === historyId ? null : historyId;
-    },
-
-    formatDate(dateStr) {
-      if (!dateStr) return 'Unknown';
-      const date = new Date(dateStr);
-      const now = new Date();
-      const diffMs = now - date;
-      const diffMins = Math.floor(diffMs / 60000);
-      const diffHours = Math.floor(diffMs / 3600000);
-      const diffDays = Math.floor(diffMs / 86400000);
-
-      if (diffMins < 1) return 'Just now';
-      if (diffMins < 60) return `${diffMins} minute${diffMins > 1 ? 's' : ''} ago`;
-      if (diffHours < 24) return `${diffHours} hour${diffHours > 1 ? 's' : ''} ago`;
-      if (diffDays < 7) return `${diffDays} day${diffDays > 1 ? 's' : ''} ago`;
-      
-      return date.toLocaleDateString() + ' ' + date.toLocaleTimeString();
-    },
-
-    formatTokenDate(dateStr) {
-      if (!dateStr) return 'Never';
-      return new Date(dateStr).toLocaleDateString();
-    },
-
-    // Computed: Flatten folders for select dropdown (includes nested with indentation)
-    get flattenedFolders() {
-      const result = [];
-      const flatten = (folders, depth = 0) => {
-        for (const folder of folders) {
-          result.push({
-            id: folder.id,
-            name: folder.name,
-            displayName: '  '.repeat(depth) + (depth > 0 ? '└ ' : '') + folder.name
-          });
-          if (folder.children && folder.children.length > 0) {
-            flatten(folder.children, depth + 1);
-          }
-        }
-      };
-      flatten(this.folders);
-      return result;
-    },
-
-    // Folder management
-    showNewFolderModal() {
-      this.editingFolder = { name: '', parent_id: '' };
-      this.showFolderModal = true;
-    },
-
-    renameFolder(folder) {
-      this.editingFolder = { id: folder.id, name: folder.name, parent_id: folder.parent_id || '' };
-      this.showFolderModal = true;
-    },
-
-    async saveFolder() {
-      if (!this.editingFolder.name.trim()) {
-        showToast('Folder name is required', 'error');
-        return;
-      }
-
-      const data = {
-        name: this.editingFolder.name,
-        parent_id: this.editingFolder.parent_id ? parseInt(this.editingFolder.parent_id) : null
-      };
-
-      let result;
-      if (this.editingFolder.id) {
-        // Update existing folder
-        result = await api.put(`/api/v1/folders/${this.editingFolder.id}`, data);
-      } else {
-        // Create new folder
-        result = await api.post('/api/v1/folders', data);
-      }
-
-      if (result && !result.error) {
-        this.showFolderModal = false;
-        await this.loadFolders();
-        showToast(this.editingFolder.id ? 'Folder renamed' : 'Folder created');
-      } else {
-        showToast(result?.error?.message || 'Failed to save folder', 'error');
-      }
-    },
-
-    async deleteFolder(folder) {
-      if (!confirm(`Delete folder "${folder.name}"? Snippets in this folder will not be deleted.`)) return;
-
-      const result = await api.delete(`/api/v1/folders/${folder.id}`);
-      if (!result || !result.error) {
-        await this.loadFolders();
-        if (this.filter.folderId === folder.id) {
-          this.clearFilters();
-        }
-        showToast('Folder deleted');
-      } else {
-        showToast('Failed to delete folder', 'error');
-      }
-    },
-
-    // Tag management
-    renameTag(tag) {
-      this.editingTag = { id: tag.id, name: tag.name };
-      this.showTagModal = true;
-    },
-
-    async saveTag() {
-      if (!this.editingTag.name.trim()) {
-        showToast('Tag name is required', 'error');
-        return;
-      }
-
-      const result = await api.put(`/api/v1/tags/${this.editingTag.id}`, {
-        name: this.editingTag.name
-      });
-
-      if (result && !result.error) {
-        this.showTagModal = false;
-        await this.loadTags();
-        showToast('Tag renamed');
-      } else {
-        showToast(result?.error?.message || 'Failed to rename tag', 'error');
-      }
-    },
-
-    async deleteTag(tag) {
-      if (!confirm(`Delete tag "${tag.name}"? This will remove the tag from all snippets.`)) return;
-
-      const result = await api.delete(`/api/v1/tags/${tag.id}`);
-      if (!result || !result.error) {
-        await this.loadTags();
-        if (this.filter.tagId === tag.id) {
-          this.clearFilters();
-        }
-        showToast('Tag deleted');
-      } else {
-        showToast('Failed to delete tag', 'error');
-      }
-    },
-
-    // Backup functions
-    async exportBackup() {
-      this.backupLoading = true;
-      try {
-        const params = new URLSearchParams({
-          format: this.backupOptions.format
-        });
-        if (this.backupOptions.password) {
-          params.append('password', this.backupOptions.password);
-        }
-
-        const response = await fetch(`/api/v1/backup/export?${params}`, {
-          credentials: 'include'
-        });
-
-        if (!response.ok) {
-          const error = await response.json();
-          throw new Error(error.error?.message || 'Export failed');
-        }
-
-        // Get filename from Content-Disposition header
-        const disposition = response.headers.get('Content-Disposition');
-        let filename = 'snipo-backup.json';
-        if (disposition) {
-          const match = disposition.match(/filename="(.+)"/);
-          if (match) filename = match[1];
-        }
-
-        // Download file
-        const blob = await response.blob();
-        const url = URL.createObjectURL(blob);
-        const a = document.createElement('a');
-        a.href = url;
-        a.download = filename;
-        document.body.appendChild(a);
-        a.click();
-        document.body.removeChild(a);
-        URL.revokeObjectURL(url);
-
-        showToast('Backup downloaded successfully');
-      } catch (err) {
-        showToast(err.message || 'Failed to export backup', 'error');
-      }
-      this.backupLoading = false;
-    },
-
-    async importBackup() {
-      if (!this.backupFile) {
-        showToast('Please select a backup file', 'error');
-        return;
-      }
-
-      this.backupLoading = true;
-      this.importResult = null;
-
-      try {
-        const formData = new FormData();
-        formData.append('file', this.backupFile);
-        formData.append('strategy', this.importOptions.strategy);
-        if (this.importOptions.password) {
-          formData.append('password', this.importOptions.password);
-        }
-
-        const response = await fetch('/api/v1/backup/import', {
-          method: 'POST',
-          credentials: 'include',
-          body: formData
-        });
-
-        const result = await response.json();
-
-        if (!response.ok) {
-          throw new Error(result.error?.message || 'Import failed');
-        }
-
-        this.importResult = result;
-        this.backupFile = null;
-
-        // Reload data
-        await Promise.all([
-          this.loadSnippets(),
-          this.loadTags(),
-          this.loadFolders()
-        ]);
-
-        showToast('Backup imported successfully');
-      } catch (err) {
-        showToast(err.message || 'Failed to import backup', 'error');
-      }
-      this.backupLoading = false;
-    },
-
-    async loadS3Status() {
-      try {
-        const result = await api.get('/api/v1/backup/s3/status');
-        if (result) {
-          this.s3Status = result;
-          if (result.enabled) {
-            await this.loadS3Backups();
-          }
-        }
-      } catch (err) {
-        console.error('Failed to load S3 status:', err);
-      }
-    },
-
-    async loadS3Backups() {
-      try {
-        const result = await api.get('/api/v1/backup/s3/list');
-        if (result && result.backups) {
-          this.s3Backups = result.backups;
-        }
-      } catch (err) {
-        console.error('Failed to load S3 backups:', err);
-      }
-    },
-
-    async syncToS3() {
-      this.backupLoading = true;
-      try {
-        const result = await api.post('/api/v1/backup/s3/sync', {
-          format: this.backupOptions.format,
-          password: this.backupOptions.password
-        });
-
-        if (result && !result.error) {
-          await this.loadS3Backups();
-          showToast('Backup synced to S3 successfully');
-        } else {
-          throw new Error(result?.error?.message || 'Sync failed');
-        }
-      } catch (err) {
-        showToast(err.message || 'Failed to sync to S3', 'error');
-      }
-      this.backupLoading = false;
-    },
-
-    async restoreFromS3(key) {
-      if (!confirm('Restore from this backup? This will import the backup data.')) return;
-
-      this.backupLoading = true;
-      try {
-        const result = await api.post('/api/v1/backup/s3/restore', {
-          key: key,
-          strategy: this.importOptions.strategy,
-          password: this.importOptions.password
-        });
-
-        if (result && !result.error) {
-          await Promise.all([
-            this.loadSnippets(),
-            this.loadTags(),
-            this.loadFolders()
-          ]);
-          showToast('Backup restored successfully');
-        } else {
-          throw new Error(result?.error?.message || 'Restore failed');
-        }
-      } catch (err) {
-        showToast(err.message || 'Failed to restore from S3', 'error');
-      }
-      this.backupLoading = false;
-    },
-
-    async deleteS3Backup(key) {
-      if (!confirm('Delete this backup from S3?')) return;
-
-      try {
-        const result = await api.delete(`/api/v1/backup/s3/delete?key=${encodeURIComponent(key)}`);
-        if (!result || !result.error) {
-          await this.loadS3Backups();
-          showToast('Backup deleted from S3');
-        } else {
-          throw new Error(result?.error?.message || 'Delete failed');
-        }
-      } catch (err) {
-        showToast(err.message || 'Failed to delete backup', 'error');
-      }
-    },
-
-    formatFileSize(bytes) {
-      if (bytes === 0) return '0 B';
-      const k = 1024;
-      const sizes = ['B', 'KB', 'MB', 'GB'];
-      const i = Math.floor(Math.log(bytes) / Math.log(k));
-      return parseFloat((bytes / Math.pow(k, i)).toFixed(1)) + ' ' + sizes[i];
-    },
-
-    formatDate(dateStr) {
-      if (!dateStr) return '';
-      return new Date(dateStr).toLocaleString();
-    }
-  }));
-
-  // Login form
-  Alpine.data('loginForm', () => ({
-    password: '',
-    error: '',
-    loading: false,
-
-    async login() {
-      this.loading = true;
-      this.error = '';
-
-      try {
-        const response = await fetch('/api/v1/auth/login', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          credentials: 'include',
-          body: JSON.stringify({ password: this.password })
-        });
-
-        const result = await response.json();
-
-        if (result.success) {
-          window.location.href = '/';
-        } else {
-          this.error = result.error?.message || 'Invalid password';
-        }
-      } catch (err) {
-        this.error = 'Connection error';
-      }
-
-      this.loading = false;
-    }
-  }));
-
-  // Public snippet view
-  Alpine.data('publicSnippet', () => ({
-    snippet: null,
-    loading: true,
-    error: false,
-    errorMessage: '',
-
-    async init() {
-      // Get snippet ID from URL path: /s/{id}
-      const path = window.location.pathname;
-      const match = path.match(/\/s\/([a-zA-Z0-9]+)/);
-
-      if (!match) {
-        this.error = true;
-        this.errorMessage = 'Invalid snippet URL';
-        this.loading = false;
-        return;
-      }
-
-      const snippetId = match[1];
-
-      try {
-        const response = await fetch(`/api/v1/snippets/public/${snippetId}`);
-        const result = await response.json();
-
-        if (response.ok && result) {
-          this.snippet = result;
-          this.$nextTick(() => {
-            if (typeof Prism !== 'undefined') {
-              Prism.highlightAll();
-            }
-          });
-        } else {
-          this.error = true;
-          this.errorMessage = result.error?.message || 'This snippet is not available or not public';
-        }
-      } catch (err) {
-        this.error = true;
-        this.errorMessage = 'Failed to load snippet';
-      }
-
-      this.loading = false;
-    },
-
-    async copyCode() {
-      if (this.snippet?.content) {
-        await navigator.clipboard.writeText(this.snippet.content);
-        showToast('Code copied to clipboard');
-      }
-    },
-
-    getLanguageColor(lang) {
-      const colors = {
-        javascript: '#f7df1e',
-        typescript: '#3178c6',
-        python: '#3776ab',
-        go: '#00add8',
-        rust: '#dea584',
-        java: '#b07219',
-        csharp: '#239120',
-        cpp: '#f34b7d',
-        cuda: '#76b900',
-        ruby: '#cc342d',
-        php: '#777bb4',
-        swift: '#fa7343',
-        kotlin: '#a97bff',
-        html: '#e34c26',
-        css: '#1572b6',
-        sql: '#e38c00',
-        bash: '#4eaa25',
-        json: '#292929',
-        yaml: '#cb171e',
-        markdown: '#083fa1',
-        plaintext: '#6b7280'
-      };
-      return colors[lang] || '#6b7280';
-    },
-
-    formatDate(dateStr) {
-      if (!dateStr) return '';
-      const date = new Date(dateStr);
-      return date.toLocaleDateString();
-    },
-
-    // Expose helpers globally to avoid Alpine scope issues
-    // Proxies to global functions for component access
-    autoResizeInput(element) { window.autoResizeInput(element); },
-    autoResizeSelect(element) { window.autoResizeSelect(element); }
-
-  }));
-});
-
-
-
-// Keyboard shortcuts
-document.addEventListener('keydown', (e) => {
-  // Ctrl/Cmd + K: Focus search
-  if ((e.ctrlKey || e.metaKey) && e.key === 'k') {
-    e.preventDefault();
-    document.querySelector('.search-input')?.focus();
-  }
-
-  // Ctrl/Cmd + N: New snippet
-  if ((e.ctrlKey || e.metaKey) && e.key === 'n') {
-    e.preventDefault();
-    const app = Alpine.$data(document.querySelector('[x-data="snippetsApp()"]'));
-    if (app) app.newSnippet();
-  }
-
-  // Escape: Close editor/modal
-  if (e.key === 'Escape') {
-    const app = Alpine.$data(document.querySelector('[x-data="snippetsApp()"]'));
-    if (app?.showEditor) app.cancelEdit();
-    if (app?.showDeleteModal) app.showDeleteModal = false;
-  }
-});
