+++ internal/web/static/js/stores/app-store.js
@@ -1,0 +1,20 @@
+// Global app store for Alpine.js
+import { theme } from '../modules/theme.js';
+
+export function initAppStore(Alpine) {
+  Alpine.store('app', {
+    sidebarOpen: window.innerWidth > 768,
+    currentView: 'snippets',
+    loading: false,
+    darkMode: theme.get() === 'dark',
+
+    toggleSidebar() {
+      this.sidebarOpen = !this.sidebarOpen;
+    },
+
+    toggleTheme() {
+      theme.toggle();
+      this.darkMode = theme.get() === 'dark';
+    }
+  });
+}
